+++ native/loudness_analyzer.h
@@ -15,7 +15,7 @@ EXPORT int loudness_analyzer_init(uint32_t channel_count,
                                   LoudnessAnalyzer** retval);
 EXPORT void loudness_analyzer_destroy(LoudnessAnalyzer* this);
 EXPORT int loudness_analyzer_add_frames(LoudnessAnalyzer* this,
-                                        int16_t* frames,
+                                        float* frames,
                                         uint32_t frame_count);
 EXPORT int loudness_analyzer_get_loudness_and_peak(LoudnessAnalyzer* this, double* gain, double* peak);
 EXPORT int loudness_analyzer_get_momentary_loudness(LoudnessAnalyzer* this, double* gain);
@@ -27,6 +27,6 @@ EXPORT int loudness_analyzer_reset(LoudnessAnalyzer* this);
 EXPORT void loudness_analyzer_apply_gain(LoudnessAnalyzer* this,
                                                    double gain_to_apply,
                                                    double previously_applied_gain,
-                                                   int16_t* frames,
+                                                   float* frames,
                                                    uint32_t frame_count);
 #endif //LOUDNESS_ANALYZER_H
